Enhancements and Corrections
The first version was working, but setting thread and process priority can have unexpected problems.  Therefore, incrementing priority can be activated via compiler macro INCREMENTPRIORITY, which will not be defined by default.
In addition, a compiler switch CONSOLEPROGRAM can be defined to generate a console program that generates some logging output that may help to analyse the behavior in case of any malfunction.
The hook has a configurable timeout for the timer function instead of a fixed timer of 45 milliseconds now. Values between 32 and 1024 can be configured now via environment variable NoEdgeTimeout.
In addition, the hook dll provides a function void SetTimerTick(DWORD) in addition to the hook function. This function can be used to pass the time component of the WM_TIMER message to the hook before processing the timer function. As result, the timer function can generate a key press event with a better time stamp.