Configuration of Process Priority Added
Now it is possible to configure the process priority via environment variable NoEdgePriority. Valid values for NoEdgePriorita are: realtime, high, abovenormal, belownormal and idle.
If this environment variable is undefined or invalid, normal priority will be used (the default). Otherwise, the corresponding priority class will be set and the thread priority will be changed to THREAD_PRIORITY_HIGHEST.